moved pkg/recovery to internal/server/middlewares
