+++ internal/version/version.go
@@ -3,5 +3,5 @@ package version
 const (
 	WDB_VERSION = "v1.1.6"
 	CLI_VERSION = "v0.0.1"
-	BUILD_DATE = "2023-03-15_10:15:30"
+	BUILD_DATE  = "2023-03-15_10:15:30"
 )
